+++ microservicio/dominio/src/main/java/com/ceiba/scotter/modelo/dto/DtoScotter.java
@@ -9,7 +9,7 @@ public class DtoScotter {
     private Long id;
     private String marca;
     private String modelo;
-    private int precio;
+    private double precio;
     private String ciudad;
     private Long vendedor;
     private String estado;
